requirements.txt
python-dotenv>=1.0.0
pytz>=2025.2
asyncio==4.0.0
requests==2.32.5
tenacity>=9.1.2
eth-account==0.13.4
ckzg>=2.0.0
httpx<0.28.0,>=0.27.0
ledgereth<0.10.0,>=0.9.1
marshmallow-dataclass<9.0.0,>=8.6.1
poseidon-py<0.2.0,>=0.1.0
pydantic<3.0.0,>=2.0.0
starknet-crypto-py<0.3.0,>=0.2.0
starknet-py<0.29.0,>=0.28.0
web3>=7.0.0,<8.0.0
websockets<16.0,>=15.0

# Paradex Python SDK (install separately with no deps)
# pip install -r requirements_paradex.txt

# Lighter exchange SDK
git+https://github.com/elliottech/lighter-python.git@d0009799970aad54ebb940aa3dc90cbc00028c54
